+++ Sources/MixpanelPrivate.h
@@ -10,7 +10,7 @@
 #import "SessionMetadata.h"
 #import "MixpanelType.h"
 #import "MixpanelPersistence.h"
-#include <TargetConditionals.h>
+#import <TargetConditionals.h>
 
 #if TARGET_OS_IOS
 #import <UIKit/UIKit.h>
@@ -86,8 +86,6 @@ static NSString *const MPDebugUsedPeopleKey = @"MPDebugUsedPeopleKey";
 @property (nonatomic, strong) NSMutableDictionary *timedEvents;
 @property (nonatomic, strong) SessionMetadata *sessionMetadata;
 
-@property (nonatomic) BOOL decideResponseCached;
-
 
 @property (nonatomic, assign) BOOL optOutStatus;
 @property (nonatomic, assign) BOOL optOutStatusNotSet;
@@ -103,7 +101,7 @@ static NSString *const MPDebugUsedPeopleKey = @"MPDebugUsedPeopleKey";
 
 - (NSString *)deviceModel;
 
-- (NSString *)defaultDistinctId;
+- (NSString *)defaultDeviceId;
 - (void)archive;
 
 // for group caching